First refactor to use addon-test-support
