+++ addon-test-support/index.js
@@ -1,0 +1,2 @@
+export { default as calendarSelect } from './calendar-select';
+export { default as calendarCenter } from './calendar-center';
